Maine Project for Data Visualization
This is my first project for Data Visualization. It focuses on the
population of each county in Maine vs. the top 20 average internet
speeds by town. In this project, there are two text files that the
program reads from in order to plot on the map: the first houses the
pixel locations and data regarding the population, and the second
handles the pixel locations and data of the internet speeds. To cycle
between the two viewing modes, the user can press 'q' or 'w'.